% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{median_calc}
\alias{median_calc}
\title{Calculate Median}
\usage{
median_calc(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The median of the vector.
}
\description{
This function calculates the median of a numeric vector.
}
\examples{
median_calc(c(1, 2, 3, 4, 5))
}
